Move mapping of dynamic options for PDF to separate service.
eases testing and reduces the amount of code in one service
